CDK/Sagemaker/requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_sagemaker
aws-cdk.aws_glue
aws-cdk.aws_iam